refactor(): further cleanups & fixes - fixed CAG & CAGMakers, moved methods/functions arround - added a few missing main exports
