+++ src/FuzzyFactory2d.js
@@ -19,18 +19,6 @@ FuzzyCAGFactory.prototype = {
     let vertex0 = this.getVertex(sourceside.vertex0)
     let vertex1 = this.getVertex(sourceside.vertex1)
     return new Side(vertex0, vertex1)
-  },
-
-  getCAG: function (sourcecag) {
-    let _this = this
-    let newsides = sourcecag.sides.map(function (side) {
-      return _this.getSide(side)
-    })
-        // remove bad sides (mostly a user input issue)
-        .filter(function (side) {
-          return side.length() > EPS
-        })
-    return fromSides(newsides)
   }
 }
